web-app/requirements.txt
numpy~=1.24.2
requests~=2.28.2
python-dotenv~=1.0.0
flask~=2.3.3